fix(ai-workflow-builder): Fix plan mode context and discovery flow
- Simplify user request tracking: replace userRequest/refinementRequest
  with userRequests array that collects ALL human messages
- Fix discoveryContext not reaching builder: add nodesFound to
  PlannerGraphAnnotation so LangGraph preserves it
- Capture bestPractices, templateIds, cachedTemplates from planner's
  tool calls (reusing Discovery pattern)
- Fix node hydration by storing nodes in state annotation
